+++ Explorer/Assets/DCL/Chat/ChatController.cs
@@ -9,6 +9,8 @@ using DCL.Chat.EventBus;
 using DCL.Diagnostics;
 using DCL.Communities;
 using DCL.Communities.CommunitiesCard;
+using DCL.Communities.CommunitiesDataProvider;
+using DCL.Communities.CommunitiesDataProvider.DTOs;
 using DCL.Friends;
 using DCL.Friends.UserBlocking;
 using DCL.Input;
