+++ Explorer/Assets/DCL/Chat/UserConnectivityInfoProvider.cs
@@ -1,6 +1,8 @@
 using Cysharp.Threading.Tasks;
 using DCL.Chat.History;
 using DCL.Communities;
+using DCL.Communities.CommunitiesDataProvider;
+using DCL.Communities.CommunitiesDataProvider.DTOs;
 using DCL.Diagnostics;
 using DCL.Utilities.Extensions;
 using Decentraland.SocialService.V2;
