+++ Explorer/Assets/DCL/Communities/CommunitiesBrowser/MyCommunityCardView.cs
@@ -1,3 +1,4 @@
+using DCL.Communities.CommunitiesDataProvider.DTOs;
 using DCL.UI;
 using System;
 using TMPro;
