+++ Explorer/Assets/DCL/Communities/CommunitiesCard/Members/FriendshipHelpers.cs
@@ -1,3 +1,4 @@
+using DCL.Communities.CommunitiesDataProvider.DTOs;
 using DCL.UI.GenericContextMenu.Controls.Configs;
 using System;
