+++ Explorer/Assets/DCL/Communities/CommunitiesDataProvider/DTOs/CommunityMemberRole.cs
@@ -2,7 +2,7 @@
 using System;
 
 // ReSharper disable InconsistentNaming
-namespace DCL.Communities
+namespace DCL.Communities.CommunitiesDataProvider.DTOs
 {
     [Serializable]
     public enum CommunityMemberRole
