+++ Explorer/Assets/DCL/Communities/CommunitiesDataProvider/DTOs/CommunityThumbnails.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace DCL.Communities
+namespace DCL.Communities.CommunitiesDataProvider.DTOs
 {
     [Serializable]
     public struct CommunityThumbnails
