+++ Explorer/Assets/DCL/Communities/CommunitiesDataProvider/DTOs/FriendshipStatus.cs
@@ -1,5 +1,5 @@
 
-namespace DCL.Communities
+namespace DCL.Communities.CommunitiesDataProvider.DTOs
 {
     public enum FriendshipStatus
     {
