+++ Explorer/Assets/DCL/Communities/CommunitiesDataProvider/DTOs/GetCommunityPlacesResponse.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace DCL.Communities
+namespace DCL.Communities.CommunitiesDataProvider.DTOs
 {
     [Serializable]
     public class GetCommunityPlacesResponse
