+++ Explorer/Assets/DCL/Communities/CommunitiesDataProvider/DTOs/GetOnlineCommunityMembersResponse.cs
@@ -1,7 +1,6 @@
-
 using System;
 
-namespace DCL.Communities
+namespace DCL.Communities.CommunitiesDataProvider.DTOs
 {
     [Serializable]
     public class GetOnlineCommunityMembersResponse
