+++ Explorer/Assets/DCL/Communities/CommunityCreation/CommunityCreationEditionView.cs
@@ -1,5 +1,6 @@
 using Cysharp.Threading.Tasks;
 using DCL.Audio;
+using DCL.Communities.CommunitiesDataProvider.DTOs;
 using DCL.UI;
 using DCL.UI.SelectorButton;
 using DCL.UI.Utilities;
