+++ Explorer/Assets/DCL/PluginSystem/Global/ChatPlugin.cs
@@ -9,6 +9,7 @@ using DCL.Chat.History;
 using DCL.Chat.MessageBus;
 using DCL.Chat.EventBus;
 using DCL.Communities;
+using DCL.Communities.CommunitiesDataProvider;
 using DCL.Friends;
 using DCL.Friends.UserBlocking;
 using DCL.FeatureFlags;
